requirements.txt
anyio==4.6.2.post1
APScheduler==3.10.4
asgiref==3.8.1
cachetools==5.5.0
certifi==2024.8.30
chardet==5.2.0
Django==5.1.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
hstspreload==2024.11.1
httpcore==1.0.7
httpx==0.27.2
hyperframe==6.0.1
idna==3.10
packaging==24.2
pip-review==1.3.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
python-telegram-bot==21.7
pytz==2024.2
rfc3986==2.0.0
setuptools==75.5.0
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.2
tornado==6.4.1
tzlocal==5.2
